Enable default Skipdata behavior
In this mode data (or invalid instructions) will be skipped and
the disassembly will attempt continue. Prior to this change, it aborts.

In the default Skipdata mode of operation, the number of bytes skipped
is based upon the architecture and mode. Skipped data is included in the
disassembly as a sequence of byte, prefixed by the Capstone
SKIPDATA_MNEM mnemonic (".byte" at the time of writing).